feat: integrate sticky display group with nav-stack; add title variant 'header'
